+++ src/Type/Types.php
@@ -24,6 +24,7 @@ final class Types
     public const FLOAT = GraphQLType::FLOAT;
     public const DATETIME = DateTimeType::DATETIME;
     public const DATE = DateType::DATE;
+    public const TIME = TimeType::TIME;
     public const ANY = AnyType::ANY;
     public const DYNAMIC_OBJECT = DynamicObjectType::DYNAMIC_OBJECT;