+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.1.3 - [Unreleased]
  * Added input type guesser to guess the correctness GraphQL type in forms fields [#10](https://github.com/ynloultratech/graphql-bundle/issues/10)
  * Removed experimental `roles` and role AuthorizationChecker.
  * Some minors bugs fixed
+ * Added `Date` type using *ISO-8601* format: 2018-06-21
 
 v1.1.2 - 2018-06-27
 ----