+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.1.3 - [Unreleased]
  * Removed experimental `roles` and role AuthorizationChecker.
  * Some minors bugs fixed
  * Added `Date` type using *ISO-8601* format: 2018-06-21
+ * Added `Time` type using *ISO-8601* format: 13:18:05
 
 v1.1.2 - 2018-06-27
 ----